addded add post page and edit post page and make to work, adding post and editing post only with logged user

+++ client/src/app/feature/posts/components/post-details/post-details.component.ts
@@ -51,10 +51,16 @@ export class PostDetailsComponent implements OnInit {
     );
   }
 
-  onDeletePost() {
-    this.postsService.deletePost(this.post().id);
+  onPostNavigate(route: 'delete' | 'edit') {
+    if (route === 'delete') {
+      this.postsService.deletePost(this.post().id);
 
-    this.router.navigate(['/']);
+      this.router.navigate(['/']);
+    }
+
+    if (route === 'edit') {
+      this.router.navigate([`edit-post/${this.post().id}`]);
+    }
   }
 
   onLoadmore() {
